學員作品/運動場地查詢/requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
Pillow==9.5.0
requests==2.29.0
urllib3==1.26.15
